% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/code_region_name.R
\name{dg_district_code}
\alias{dg_district_code}
\title{code district name in daegu.}
\usage{
dg_district_code(
  data,
  new_var = "district",
  region_code_var = "d_num",
  dong_name_var = "d",
  county_code = "'27"
)
}
\arguments{
\item{data}{data set.}

\item{new_var}{variable created.}

\item{region_code_var}{variable of region code.}

\item{dong_name_var}{variable name in Korean of small region.}

\item{county_code}{region code of Daegu in population data.}
}
\value{
data including a new district varible.
}
\description{
dg_district_code function will code district name of Daege.
}
